opencti-worker/requirements.txt
requests
PyYAML
grakn-client
elasticsearch
pika